refactor(utm): migrate UTM cookie handling to utmService and remove deprecated action files
