+++ web-app/src/lib/actions/index.ts
@@ -5,4 +5,3 @@ export * from "./invitation";
 export * from "./job";
 export * from "./organization";
 export * from "./types";
-export * from "./utm";
