--- web-app/src/lib/actions/utm/index.ts
@@ -1,1 +1,0 @@
-export * from "./action";
